refactor(native-rd/i18n): split lintSource CLI from pure module
Moves the file-system bootstrap (main(), import.meta.dir, import.meta.main)
out of lintSource.ts into a thin lintSource.cli.ts wrapper. The pure module
now exports lintEnDir() so tests can import the linter without parsing
import.meta — babel-preset-expo targets Hermes and does not enable
unstable_transformImportMeta, which broke jest test discovery.

Also adds "bun" types to tsconfig.scripts-test.json so the test config
resolves node:fs/node:path transitively imported by lintSource.ts. Refs #158.
