+++ apps/native-rd/tsconfig.scripts-test.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "types": ["jest"]
+    "types": ["jest", "bun"]
   },
   "include": ["scripts/**/__tests__/**/*.ts"],
   "exclude": []
